externalize const and interfaces
